+++ slurm/slurmdb.h
@@ -300,7 +300,7 @@ typedef struct {
 typedef struct {
 	double act_cpufreq;	/* contains actual average cpu frequency */
 	double cpu_ave;
-	double consumed_energy; /* contains energy consumption in joules */
+	uint64_t consumed_energy; /* contains energy consumption in joules */
 	uint32_t cpu_min;
 	uint32_t cpu_min_nodeid; /* contains which node number it was on */
 	uint32_t cpu_min_taskid; /* contains which task number it was on */
